pyproject.toml
[tool.poetry]
name = "transactions-test-task"
version = "0.1.0"
description = ""
authors = ["dkarpushkin <dmitry.karpushkin.true@gmail.com>"]
readme = "README.md"
packages = [{include = "transactions_test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
uvloop = "^0.17.0"
aiohttp = "^3.8.4"
psycopg2 = "^2.9.5"
requests = "^2.28.2"
pytest = "7.2.2"
environs = "^9.5.0"
gino = "^1.0.1"
alembic = "^1.10.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
